refactor: move submissionslist.tsx complex logic to a custom hook
