+++ src/app/components/submissions-list/SubmissionsList.tsx
@@ -1,10 +1,6 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 'use client';
-import { usePathname, useRouter, useSearchParams } from 'next/navigation';
-import { useCallback, useEffect, useState } from 'react';
-import { useFilters } from '../../../lib/state/FiltersContext';
-import { PageSize, usePagination } from '../../../lib/state/PaginationContext';
-import { useShouldUpdate } from '../../../lib/state/ShouldUpdateContext';
+import { useState } from 'react';
 import { SUBMISSIONS_LIST_SELECTORS } from '../../../lib/test-selectors/components/submissions-list.selectors';
 import { PostFilters } from '../../posts/page';
 import Empty from '../empty/Empty';
@@ -13,15 +9,10 @@ import { Filter } from '../filter-bar/FilterBar';
 import Loader from '../loader/Loader';
 import Pagination from '../pagination/Pagination';
 import { DeleteSubmissionForm } from '../submission-forms/delete-submission-form/DeleteSubmissionForm';
-import { Submission } from '../submission-forms/schema';
 import { TagLink } from '../tag-link/TagLink';
 import { ReplyForm } from '../thread/ReplyForm';
-import {
-  getSubmissionsAction,
-  GetSubmissionsActionArguments,
-  PaginatedResponse
-} from './actions';
 import './SubmissionsList.css';
+import { useSubmissionsList } from './use-submissions-list';
 
 export default function SubmissionsList({
   contextId = 'default',
@@ -36,209 +27,19 @@ export default function SubmissionsList({
   onlyMine?: boolean;
   filters?: Filter<PostFilters>[];
 }) {
-  const router = useRouter();
-  const pathName = usePathname();
-  const searchParams = useSearchParams();
-  const tagSearchParams = searchParams.get('tags');
-  const pageSearchParam = searchParams.get('page');
-  const { state: filtersState, dispatch: dispatchFilters } = useFilters();
-  const { state: shouldUpdate, dispatch: dispatchShouldUpdate } =
-    useShouldUpdate();
-  const { state: paginationState, dispatch: dispatchPagination } =
-    usePagination();
-  const paginationContext = paginationState[contextId];
-  const filtersContext = filtersState[contextId];
-  const tagSearchParamFilters: Filter<PostFilters>[] = tagSearchParams
-    ? [{ name: 'tags', value: tagSearchParams }]
-    : [];
-
-  // TODO: https://github.com/vercel/next.js/issues/65673#issuecomment-2112746191
-  // const [state, formAction] = useActionState(getSubmissionsAction, initialState)
-  const [loading, setLoading] = useState(true);
-  const [response, setResponse] = useState<
-    | {
-        data?: PaginatedResponse<Submission>;
-        message?: string;
-        error?: string;
-      }
-    | undefined
-  >();
   const [activeThreadId, setActiveThreadId] = useState<number | null>(null);
-
-  const getArgs = useCallback(() => {
-    // prioritize url params, context state data, and then defer to prop data
-    const data: {
-      currentPage: number;
-      onlyMine: boolean;
-      providerAccountId: string;
-      filters: Filter[];
-      pageSize: number;
-    } = {
-      currentPage:
-        Number(pageSearchParam) || paginationContext?.currentPage || page,
-      filters: tagSearchParamFilters || filtersContext?.filters || filters,
-      onlyMine,
-      providerAccountId,
-      pageSize: paginationContext?.pageSize || PageSize.TEN
-    };
-
-    return data;
-  }, [
-    paginationContext,
-    page,
-    tagSearchParams,
-    pageSearchParam,
-    filters,
-    onlyMine,
-    providerAccountId
-  ]);
-
-  // TODO: extract fetch submissions and useEffects into custom hook
-
-  const fetchSubmissions = useCallback(
-    async (args: GetSubmissionsActionArguments) => {
-      setLoading(true);
-      setResponse(await getSubmissionsAction(args));
-      setLoading(false);
-    },
-    []
-  );
-
-  /** response listener/handler */
-  useEffect(() => {
-    if (response && response.data?.pagination) {
-      const newTotalPages =
-        response.data.pagination.totalRecords /
-        (response.data.pagination.pageSize || 10);
-      const totalPages = Math.ceil(newTotalPages);
-      const requestedPage = response?.data?.pagination.currentPage;
-
-      if (requestedPage > totalPages) {
-        // replace requested page with last available page
-        dispatchPagination({
-          type: 'SET_CURRENT_PAGE',
-          payload: {
-            id: contextId,
-            currentPage: totalPages
-          }
-        });
-      }
-
-      if (requestedPage < 1) {
-        // replace requested page with first page
-        dispatchPagination({
-          type: 'SET_CURRENT_PAGE',
-          payload: {
-            id: contextId,
-            currentPage: 1
-          }
-        });
-      }
-
-      dispatchPagination({
-        type: 'SET_TOTAL_PAGES',
-        payload: {
-          id: contextId,
-          totalPages
-        }
-      });
-    }
-  }, [
-    dispatchPagination,
-    response?.data?.pagination.totalRecords,
-    response?.data?.pagination.currentPage,
-    response?.data?.pagination.pageSize,
-    contextId
-  ]);
-
-  /**
-   * shouldUpdate listener + handler
-   * triggers a refetch of latest data with latest filters/pagination withi initial page size of 10
-   * */
-  useEffect(() => {
-    if (shouldUpdate) {
-      dispatchShouldUpdate({ type: 'SET_SHOULD_UPDATE', payload: false });
-
-      fetchSubmissions(getArgs());
-    }
-  }, [shouldUpdate]);
-
-  /**
-   * context listeners: filter context filters, pagination context current page
-   *
-   * will update the url search params to match latest context data
-   */
-  useEffect(() => {
-    const latestFilters = filtersContext?.filters.find(
-      (filter) => filter.name === 'tags'
-    )?.value;
-    const params = new URLSearchParams();
-
-    if (latestFilters) {
-      params.set('tags', latestFilters);
-    }
-
-    if (page || page === 0) {
-      const maxPage = Number(paginationContext?.totalPages);
-      let newPage = paginationContext?.currentPage?.toString() || '1';
-
-      if (Number(newPage) > maxPage) {
-        newPage = maxPage.toString();
-      }
-      if (Number(newPage) < 1) {
-        newPage = '1';
-      }
-
-      // update url param page to the current pagination state value
-      params.set('page', newPage);
-    }
-
-    const newRoute = `${pathName}?${params.toString()}`;
-    router.push(newRoute);
-  }, [paginationContext?.currentPage, filtersContext?.filters]);
-
-  // page size listener
-  useEffect(() => {
-    // manual fetch so that default page size of 10 is not used from shouldUpdate dispatch handler
-    dispatchShouldUpdate({ payload: true, type: 'SET_SHOULD_UPDATE' });
-  }, [paginationContext?.pageSize]);
-
-  // route params listener: tags & current page
-  useEffect(() => {
-    // ensure context state is updated on manual url change (i.e. browser navigation action)
-    if (Number(pageSearchParam) !== paginationContext?.currentPage) {
-      dispatchPagination({
-        type: 'SET_CURRENT_PAGE',
-        payload: {
-          id: contextId,
-          currentPage: Number(pageSearchParam)
-        }
-      });
-    }
-
-    if (
-      tagSearchParams !==
-      filtersContext?.filters.find((filter) => filter.name === 'tags')?.value
-    ) {
-      dispatchFilters({
-        type: 'SET_CURRENT_FILTERS',
-        payload: {
-          id: contextId,
-          filters: tagSearchParamFilters
-        }
-      });
-    }
-
-    dispatchShouldUpdate({ type: 'SET_SHOULD_UPDATE', payload: true });
-  }, [tagSearchParams?.length, pageSearchParam]);
-
-  const isAuthorized = (authorId: string) => {
-    return providerAccountId === authorId;
-  };
 
   const toggleReplyForm = (submissionId: number) => {
     setActiveThreadId(activeThreadId === submissionId ? null : submissionId);
   };
+
+  const { isAuthorized, loading, response } = useSubmissionsList(
+    contextId,
+    providerAccountId,
+    onlyMine,
+    page,
+    filters
+  );
 
   return (
     <article
